Reconfigured command locations to the standard
