+++ src/main/java/frc/robot/subsystems/climb/climbCommands/climbTuning.java
@@ -1,5 +1,5 @@
 /* (C) RoboLancers 2026 */
-package frc.robot.commands.climbCommands;
+package frc.robot.subsystems.climb.climbCommands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.climb.Climb;
