+++ src/main/java/frc/robot/subsystems/climb/climbCommands/goToAngle.java
@@ -1,19 +1,19 @@
 /* (C) RoboLancers 2026 */
-package frc.robot.commands.climbCommands;
+package frc.robot.subsystems.climb.climbCommands;
 
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.climb.Climb;
 
-public class goToPivotAngle extends Command {
+public class goToAngle extends Command {
 
   Climb climb;
 
-  public goToPivotAngle(Climb climb) {
+  public goToAngle(Climb climb) {
     this.climb = climb;
   }
 
   public void execute(Angle angle) {
-    climb.goToPivotAngle(angle);
+    climb.goToAngle(angle);
   }
 }
